Add support for more file extensions with QuickLook
