+++ Themis/Views/Assessment/File Upload Exercise/FileRendererFactory.swift
@@ -12,9 +12,11 @@ enum FileRendererFactory {
     static func renderer(for fileExtension: FileUploadExerciseFileExtension, at url: URL) -> some View {
         switch fileExtension {
         case .pdf:
+            // We are not using QuickLook here because it allows sketching on the pages
+            // This can make the user think that the sketches will be visible to the student as well
             PDFRenderer(url: url)
-        case .png, .jpeg:
-            ImageFileRenderer(url: url)
+        case .doc, .docx, .xlsx, .csv, .txt, .png, .jpeg:
+            QuickLookFileRenderer(url: url)
         default:
             UnsupportedFileView(url: url, fileExtension: fileExtension)
         }
@@ -22,5 +24,10 @@ enum FileRendererFactory {
 }
 
 enum FileUploadExerciseFileExtension: String {
-    case pdf, jpeg, png, other
+    case pdf, jpeg, png, doc, docx, xlsx, txt, csv, other
+}
+
+/// Intended for views that can be returned by FileRendererFactory
+protocol FileRenderer {
+    var url: URL { get }
 }
